Changed type of property Expiration in CertificateCredentials from int to long.

+++ src/VaultSharp/V1/SecretsEngines/PKI/CertificateCredentials.cs
@@ -32,6 +32,6 @@ namespace VaultSharp.V1.SecretsEngines.PKI
         /// The expiration.
         /// </value>
         [JsonProperty("expiration")]
-        public int Expiration { get; set; }
+        public long Expiration { get; set; }
     }
 }
